Deleted TestHelper.cs and moved necessary code into existing TransactionBuilder.cs
